[SECURITY-904] Fold tests back into normal files
