+++ test/src/test/resources/hudson/model/DirectoryBrowserSupportTest/insideWorkspaceStructure.bat
