+++ test/src/test/resources/hudson/model/DirectoryBrowserSupportTest/outsideWorkspaceStructure.bat
