+++ test/src/test/resources/hudson/model/DirectoryBrowserSupportTest/outsideWorkspaceStructure.sh
