+++ test/src/test/resources/hudson/model/DirectoryBrowserSupportTest/outsideWorkspaceStructureWithJunctions.bat
